Separate server-system-device to fix circular dep
