+++ eshtek/server-system-device-schema.ts
@@ -1,0 +1,30 @@
+// Generated by ts-to-zod
+import { z } from "zod";
+
+const vmPassthroughDeviceChoiceSchema = z.any();
+
+const vmUsbPassthroughDeviceChoiceSchema = z.any();
+
+export const serverSystemDeviceSchema = z.object({
+  name: z.string(),
+  guid: z.string(),
+  data: z.union([
+    vmPassthroughDeviceChoiceSchema,
+    vmUsbPassthroughDeviceChoiceSchema,
+  ]),
+});
+
+export const serverSystemGPUSchema = serverSystemDeviceSchema;
+
+export const serverSystemAudioSchema = serverSystemDeviceSchema;
+
+export const serverSystemUSBSchema = serverSystemDeviceSchema;
+
+export const serverSystemPCISchema = serverSystemDeviceSchema;
+
+export const serverSystemDevicesSchema = z.object({
+  gpu: z.array(serverSystemGPUSchema).optional(),
+  audio: z.array(serverSystemAudioSchema).optional(),
+  usb: z.array(serverSystemUSBSchema).optional(),
+  pci: z.array(serverSystemPCISchema).optional(),
+});
