+++ eshtek/server.ts
@@ -4,7 +4,6 @@ import type { FileType } from '../truenas/webui/enums/file-type.enum';
 import type { NetworkInterfaceType } from '../truenas/webui/enums/network-interface.enum';
 import type { PoolStatus } from '../truenas/webui/enums/pool-status.enum';
 import type { TopologyItemStatus } from '../truenas/webui/enums/vdev-status.enum';
-import type { VmPassthroughDeviceChoice, VmUsbPassthroughDeviceChoice } from '../truenas/webui/interfaces/vm-device.interface';
 import type { AppsHealth } from './apps';
 import type { VMSHealth } from './vms';
 
@@ -309,24 +308,6 @@ export type ServerSystemData =
     | ServerSystemDataApplications
     | ServerSystemDataVirtualization
     | ServerSystemDataEmpty;
-
-export interface ServerSystemDevice {
-    name: string;
-    guid: string;
-    data: VmPassthroughDeviceChoice | VmUsbPassthroughDeviceChoice;
-}
-
-export interface ServerSystemGPU extends ServerSystemDevice {}
-export interface ServerSystemAudio extends ServerSystemDevice {}
-export interface ServerSystemUSB extends ServerSystemDevice {}
-export interface ServerSystemPCI extends ServerSystemDevice {}
-
-export interface ServerSystemDevices {
-    gpu?: ServerSystemGPU[];
-    audio?: ServerSystemAudio[];
-    usb?: ServerSystemUSB[];
-    pci?: ServerSystemPCI[];
-}
 
 export interface ServerSystem extends ServerStatusBasics {
     type: ServerStatusType.SYSTEM_OVERVIEW;
