+++ ts-to-zod.config.js
@@ -10,6 +10,11 @@ module.exports = [
         output: "eshtek/server-schema.ts",
     },
     {
+        name: "server-system-device",
+        input: "eshtek/server-system-device.ts",
+        output: "eshtek/server-system-device-schema.ts",
+    },
+    {
         name: "user",
         input: "eshtek/user.ts",
         output: "eshtek/user-schema.ts",
